refactor(theme): migrate and refactor type `OverrideComponent` by `@ant-design/cssinjs-utils`
